GOT RID OF DownloadPreprocessor class and moved its functions into IntegerPatternProcessor
